Provide Cloudformation template for deploying the requeueing Lambda function
Right now it's just stand alone (with the required IAM policy) but tempted to also provide another template with the 2 queues in it as well.
Still need to provide a scheduled event for firing the Lambda function regularly too.

Not overly sure how I feel about Cloudformation and some of the nested list stuff around Join might be better off writing in [] format instead of YAML style.

Also considering ditching the sending to S3 and instead have Cloudformation simply use the zip directly.